Make sure trajectory chunks exist prior to trajectory reconstruction.

+++ src/somd2/runner/_base.py
@@ -1646,15 +1646,17 @@ class RunnerBase:
                         _copyfile(traj_filename, f"{traj_filename}.prev")
                         traj_chunks = [f"{traj_filename}.prev"] + traj_chunks
 
-                # Load the topology and chunked trajectory files.
-                mols = _sr.load([topology0] + traj_chunks)
+                # Make sure there are trajectory chunks to process.
+                if len(traj_chunks) > 0:
+                    # Load the topology and chunked trajectory files.
+                    mols = _sr.load([topology0] + traj_chunks)
 
-                # Save the final trajectory to a single file.
-                _sr.save(mols.trajectory(), traj_filename, format=["DCD"])
+                    # Save the final trajectory to a single file.
+                    _sr.save(mols.trajectory(), traj_filename, format=["DCD"])
 
-                # Now remove the chunked trajectory files.
-                for chunk in traj_chunks:
-                    _Path(chunk).unlink()
+                    # Now remove the chunked trajectory files.
+                    for chunk in traj_chunks:
+                        _Path(chunk).unlink()
 
             # Add config and lambda value to the system properties.
             system.set_property("config", self._config.as_dict(sire_compatible=True))
